adk-service/pyproject.toml
[project]
name = "adk-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
